move typescript input gathering to project-wide util function
